mysite/requirements.txt
poetry==1.7.1
django-bootstrap4==23.2
bootstrap4==0.1.0